moving exceptions inside the errors module
