+++ lib/errors/exceptions.js
