+++ lib/utils.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var exceptions = require('./exceptions');
+var exceptions = require('./errors/exceptions');
 
 
 /**
