+++ tests/test_model_task.js
@@ -1,4 +1,4 @@
-var exceptions = require('../lib/exceptions');
+var exceptions = require('../lib/errors/exceptions');
 var Task = require('../lib/models/task');
 
 describe('Task model', function() {
